srv_pvt_api/requirements.txt
asgiref==3.3.1
azure-core==1.11.0
azure-storage-blob==12.7.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
Django==2.1.15
django-cors-headers==3.4.0
django-pyodbc==1.1.3
django-pyodbc-azure==2.1.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
pycparser==2.20
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
